+++ contracts/ibc-reflect/tests/integration.rs
@@ -93,11 +93,13 @@ fn connect(
     let id = res.messages[0].id;
 
     // fake a reply and ensure this works
+    #[allow(deprecated)]
     let response = Reply {
         id,
         gas_used: 1234567,
         result: SubMsgResult::Ok(SubMsgResponse {
             events: fake_events(&account),
+            msg_responses: vec![],
             data: None,
         }),
     };
@@ -172,11 +174,13 @@ fn proper_handshake_flow() {
     assert_eq!(0, res.accounts.len());
 
     // we get the callback from reflect
+    #[allow(deprecated)]
     let response = Reply {
         id,
         gas_used: 1234567,
         result: SubMsgResult::Ok(SubMsgResponse {
             events: fake_events(REFLECT_ADDR),
+            msg_responses: vec![],
             data: None,
         }),
     };