+++ contracts/reflect/src/contract.rs
@@ -435,9 +435,11 @@ mod tests {
         let data = Binary::from(b"foobar");
         let events = vec![Event::new("message").add_attribute("signer", "caller-addr")];
         let gas_used = 1234567u64;
+        #[allow(deprecated)]
         let result = SubMsgResult::Ok(SubMsgResponse {
             events: events.clone(),
             data: Some(data.clone()),
+            msg_responses: vec![],
         });
         let subcall = Reply {
             id,
@@ -460,7 +462,10 @@ mod tests {
         let qres: Reply = from_json(raw).unwrap();
         assert_eq!(qres.id, id);
         let result = qres.result.unwrap();
-        assert_eq!(result.data, Some(data));
+        #[allow(deprecated)]
+        {
+            assert_eq!(result.data, Some(data));
+        }
         assert_eq!(result.events, events);
     }
 }