+++ packages/std/src/results/submessages.rs
@@ -123,22 +123,27 @@ pub struct Reply {
 /// Success:
 ///
 /// ```
-/// # use cosmwasm_std::{to_vec, Binary, Event, SubMsgResponse, SubMsgResult};
+/// # use cosmwasm_std::{to_json_string, Binary, Event, SubMsgResponse, SubMsgResult};
+/// #[allow(deprecated)]
 /// let response = SubMsgResponse {
 ///     data: Some(Binary::from_base64("MTIzCg==").unwrap()),
 ///     events: vec![Event::new("wasm").add_attribute("fo", "ba")],
+///     msg_responses: vec![],
 /// };
 /// let result: SubMsgResult = SubMsgResult::Ok(response);
-/// assert_eq!(to_vec(&result).unwrap(), br#"{"ok":{"events":[{"type":"wasm","attributes":[{"key":"fo","value":"ba"}]}],"data":"MTIzCg=="}}"#);
+/// assert_eq!(
+///     to_json_string(&result).unwrap(),
+///     r#"{"ok":{"events":[{"type":"wasm","attributes":[{"key":"fo","value":"ba"}]}],"data":"MTIzCg==","msg_responses":[]}}"#,
+/// );
 /// ```
 ///
 /// Failure:
 ///
 /// ```
-/// # use cosmwasm_std::{to_vec, SubMsgResult, Response};
+/// # use cosmwasm_std::{to_json_string, SubMsgResult, Response};
 /// let error_msg = String::from("Something went wrong");
 /// let result = SubMsgResult::Err(error_msg);
-/// assert_eq!(to_vec(&result).unwrap(), br#"{"error":"Something went wrong"}"#);
+/// assert_eq!(to_json_string(&result).unwrap(), r#"{"error":"Something went wrong"}"#);
 /// ```
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
@@ -199,10 +204,20 @@ impl From<SubMsgResult> for Result<SubMsgResponse, String> {
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct SubMsgResponse {
     pub events: Vec<Event>,
+    #[deprecated = "Deprecated in the Cosmos SDK in favor of msg_responses. If your chain is running on CosmWasm 2.0 or higher, msg_responses will be filled. For older versions, the data field is still needed since msg_responses is empty in those cases."]
     pub data: Option<Binary>,
+    #[serde(default)]
+    pub msg_responses: Vec<MsgResponse>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
+pub struct MsgResponse {
+    pub type_url: String,
+    pub value: Binary,
 }
 
 #[cfg(test)]
+#[allow(deprecated)]
 mod tests {
     use super::*;
     use crate::{from_json, to_json_vec, StdError, StdResult};
@@ -211,20 +226,26 @@ mod tests {
     fn sub_msg_result_serialization_works() {
         let result = SubMsgResult::Ok(SubMsgResponse {
             data: None,
+            msg_responses: vec![],
             events: vec![],
         });
         assert_eq!(
             &to_json_vec(&result).unwrap(),
-            br#"{"ok":{"events":[],"data":null}}"#
+            br#"{"ok":{"events":[],"data":null,"msg_responses":[]}}"#
         );
 
         let result = SubMsgResult::Ok(SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
+            msg_responses: vec![MsgResponse {
+                type_url: "URL".to_string(),
+                value: Binary::from_base64("MTIzCg==").unwrap(),
+            }],
             events: vec![Event::new("wasm").add_attribute("fo", "ba")],
         });
+        println!("{}", &crate::to_json_string(&result).unwrap());
         assert_eq!(
             &to_json_vec(&result).unwrap(),
-            br#"{"ok":{"events":[{"type":"wasm","attributes":[{"key":"fo","value":"ba"}]}],"data":"MTIzCg=="}}"#
+            br#"{"ok":{"events":[{"type":"wasm","attributes":[{"key":"fo","value":"ba"}]}],"data":"MTIzCg==","msg_responses":[{"type_url":"URL","value":"MTIzCg=="}]}}"#
         );
 
         let result: SubMsgResult = SubMsgResult::Err("broken".to_string());
@@ -233,21 +254,40 @@ mod tests {
 
     #[test]
     fn sub_msg_result_deserialization_works() {
-        let result: SubMsgResult = from_json(br#"{"ok":{"events":[],"data":null}}"#).unwrap();
+        // should work without `msg_responses`
+        let result: SubMsgResult = from_json(br#"{"ok":{"events":[]}}"#).unwrap();
         assert_eq!(
             result,
             SubMsgResult::Ok(SubMsgResponse {
                 events: vec![],
                 data: None,
+                msg_responses: vec![]
+            })
+        );
+
+        // should work with `data` and no `msg_responses`
+        // this is the case for pre-2.0 CosmWasm chains
+        let result: SubMsgResult = from_json(br#"{"ok":{"events":[],"data":"aGk="}}"#).unwrap();
+        assert_eq!(
+            result,
+            SubMsgResult::Ok(SubMsgResponse {
+                events: vec![],
+                data: Some(Binary::from_base64("aGk=").unwrap()),
+                msg_responses: vec![]
             })
         );
 
         let result: SubMsgResult = from_json(
-            br#"{"ok":{"events":[{"type":"wasm","attributes":[{"key":"fo","value":"ba"}]}],"data":"MTIzCg=="}}"#).unwrap();
+            br#"{"ok":{"events":[{"type":"wasm","attributes":[{"key":"fo","value":"ba"}]}],"data":"MTIzCg==",
+            "msg_responses":[{"type_url":"URL","value":"MTIzCg=="}]}}"#).unwrap();
         assert_eq!(
             result,
             SubMsgResult::Ok(SubMsgResponse {
                 data: Some(Binary::from_base64("MTIzCg==").unwrap()),
+                msg_responses: vec![MsgResponse {
+                    type_url: "URL".to_string(),
+                    value: Binary::from_base64("MTIzCg==").unwrap(),
+                }],
                 events: vec![Event::new("wasm").add_attribute("fo", "ba")],
             })
         );
@@ -272,6 +312,10 @@ mod tests {
     fn sub_msg_result_unwrap_works() {
         let response = SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
+            msg_responses: vec![MsgResponse {
+                type_url: "URL".to_string(),
+                value: Binary::from_base64("MTIzCg==").unwrap(),
+            }],
             events: vec![Event::new("wasm").add_attribute("fo", "ba")],
         };
         let success = SubMsgResult::Ok(response.clone());
@@ -297,6 +341,7 @@ mod tests {
         let response = SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
             events: vec![Event::new("wasm").add_attribute("fo", "ba")],
+            msg_responses: vec![],
         };
         let success = SubMsgResult::Ok(response);
         let _ = success.unwrap_err();
@@ -307,6 +352,7 @@ mod tests {
         let success = SubMsgResult::Ok(SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
             events: vec![Event::new("wasm").add_attribute("fo", "ba")],
+            msg_responses: vec![],
         });
         let failure = SubMsgResult::Err("broken".to_string());
         assert!(success.is_ok());
@@ -318,6 +364,7 @@ mod tests {
         let success = SubMsgResult::Ok(SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
             events: vec![Event::new("wasm").add_attribute("fo", "ba")],
+            msg_responses: vec![],
         });
         let failure = SubMsgResult::Err("broken".to_string());
         assert!(failure.is_err());
@@ -329,6 +376,7 @@ mod tests {
         let original: Result<SubMsgResponse, StdError> = Ok(SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
             events: vec![],
+            msg_responses: vec![],
         });
         let converted: SubMsgResult = original.into();
         assert_eq!(
@@ -336,6 +384,7 @@ mod tests {
             SubMsgResult::Ok(SubMsgResponse {
                 data: Some(Binary::from_base64("MTIzCg==").unwrap()),
                 events: vec![],
+                msg_responses: vec![],
             })
         );
 
@@ -352,6 +401,7 @@ mod tests {
         let original = SubMsgResult::Ok(SubMsgResponse {
             data: Some(Binary::from_base64("MTIzCg==").unwrap()),
             events: vec![],
+            msg_responses: vec![],
         });
         let converted: Result<SubMsgResponse, String> = original.into();
         assert_eq!(
@@ -359,6 +409,7 @@ mod tests {
             Ok(SubMsgResponse {
                 data: Some(Binary::from_base64("MTIzCg==").unwrap()),
                 events: vec![],
+                msg_responses: vec![],
             })
         );